requirements.txt
PyMuPDF==1.24.14
Pillow==10.2.0
pytesseract==0.3.10
opencv-python==4.8.1.78